gold-tier/instagram/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=0.9.0
watchdog>=3.0.0
schedule>=1.2.0
